requirements.txt
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
audioread==3.0.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
fonttools==4.42.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.7
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.1
redis==5.0.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
spotipy==2.23.0
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
zipp==3.16.2
